Miscellanea commands are now available DMing the bot
