+++ src/Commands/BaseCommand.ts
@@ -1,19 +1,4 @@
-import {Command, CommandInfo, CommandoClient} from "discord.js-commando"
+import {Command} from "discord.js-commando"
 
-import {GuildResolvable} from "discord.js"
-import utils from '../Utilities/utils'
 
-export default abstract class BaseCommand extends Command {
-
-    public guildOnly: boolean = true;
-
-    isEnabledIn(guild : GuildResolvable) {
-        const isGuildOnlyDev = utils.isGuildOnlyDev(guild)
-
-        return (
-            utils.isDevelopment() && isGuildOnlyDev ||
-            utils.isProduction() && ! isGuildOnlyDev
-        )
-    }
-
-};
+export default abstract class BaseCommand extends Command {};
