+++ src/Commands/match/leave.ts
@@ -1,8 +1,8 @@
 import {CommandMessage, CommandoClient} from "discord.js-commando"
 import Match from "../../Models/Match"
-import BaseCommand from '../BaseCommand'
+import GuildCommand from '../GuildCommand'
 
-export default class JoinCommand extends BaseCommand {
+export default class JoinCommand extends GuildCommand {
     constructor(client : CommandoClient) {
         super(client, {
             name: 'leave',
