+++ src/Commands/match/list.ts
@@ -2,9 +2,9 @@ import {CommandMessage, CommandoClient} from "discord.js-commando"
 import Match from "../../Models/Match"
 import ListPendingMatches from "../../Listeners/ListPendingMatches"
 import {DMChannel, GroupDMChannel, TextChannel} from "discord.js"
-import BaseCommand from '../BaseCommand'
+import GuildCommand from '../GuildCommand'
 
-export default class ListCommand extends BaseCommand {
+export default class ListCommand extends GuildCommand {
     constructor(client : CommandoClient) {
         super(client, {
             name: 'list',
